fix es missing mapping fallback alias check
