+++ pkg/unify-query/tsdb/elasticsearch/instance.go
@@ -28,7 +28,6 @@ import (
 	"github.com/samber/lo"
 
 	"github.com/TencentBlueKing/bkmonitor-datalink/pkg/unify-query/internal/function"
-	"github.com/TencentBlueKing/bkmonitor-datalink/pkg/unify-query/log"
 	"github.com/TencentBlueKing/bkmonitor-datalink/pkg/unify-query/metadata"
 	"github.com/TencentBlueKing/bkmonitor-datalink/pkg/unify-query/metric"
 	"github.com/TencentBlueKing/bkmonitor-datalink/pkg/unify-query/query/structured"
@@ -406,7 +405,7 @@ func (i *Instance) esQuery(ctx context.Context, qo *queryOption, fact *FormatFac
 	if fallbackRes, ok := i.tryFallbackEmptyMissingMappingIndexes(ctx, span, client, qo, source, countQuery, err, res); ok {
 		res = fallbackRes
 		err = nil
-		recordESQueryShards(ctx, span, qo, res)
+		recordESQueryShardsWithPrefix(ctx, span, qo, res, "fallback_retry_")
 	}
 	if err = handleESError(ctx, qo.conn.Address, err, res); err != nil {
 		return nil, err
@@ -424,354 +423,6 @@ func (i *Instance) esQuery(ctx context.Context, qo *queryOption, fact *FormatFac
 		ctx, queryCost, metadata.ElasticsearchStorageType, qo.conn.Address,
 	)
 	return res, err
-}
-
-const (
-	esShardFailureSampleLimit     = 3
-	esShardFailureReasonMaxLength = 512
-	esMissingMappingReasonPrefix  = "No mapping found for ["
-	esMissingMappingReasonSuffix  = "] in order to sort on"
-)
-
-type esShardFailureSample struct {
-	Shard  int    `json:"shard"`
-	Index  string `json:"index"`
-	Status string `json:"status,omitempty"`
-	Reason string `json:"reason,omitempty"`
-}
-
-type esMissingMappingFailure struct {
-	Index string
-	Field string
-}
-
-func recordESQueryShards(ctx context.Context, span *trace.Span, qo *queryOption, res *elastic.SearchResult) {
-	if res == nil {
-		return
-	}
-
-	span.Set("timed_out", res.TimedOut)
-	if res.Shards == nil {
-		if res.TimedOut {
-			span.Set("shards_failures_count", 0)
-			log.Warnf(ctx, "es query shard abnormal index: %+v, timed_out: %v, shards info is nil", esQueryIndexes(qo), res.TimedOut)
-		}
-		return
-	}
-
-	span.Set("shards_total", res.Shards.Total)
-	span.Set("shards_successful", res.Shards.Successful)
-	span.Set("shards_failed", res.Shards.Failed)
-	span.Set("shards_skipped", res.Shards.Skipped)
-
-	if res.Shards.Failed <= 0 && !res.TimedOut {
-		return
-	}
-
-	failuresCount := countESShardFailures(res.Shards.Failures)
-	span.Set("shards_failures_count", failuresCount)
-	failuresSample := buildESShardFailureSample(res.Shards.Failures)
-	failuresSampleJson := marshalESShardFailureSample(failuresSample)
-	if len(failuresSample) > 0 {
-		span.Set("shards_failures_sample", failuresSampleJson)
-	}
-
-	log.Warnf(
-		ctx,
-		"es query shard abnormal index: %+v, timed_out: %v, shards_total: %d, shards_successful: %d, shards_failed: %d, shards_skipped: %d, failures_count: %d, failures_sample: %s",
-		esQueryIndexes(qo), res.TimedOut, res.Shards.Total, res.Shards.Successful, res.Shards.Failed, res.Shards.Skipped, failuresCount, failuresSampleJson,
-	)
-}
-
-func (i *Instance) tryFallbackEmptyMissingMappingIndexes(
-	ctx context.Context,
-	span *trace.Span,
-	client *elastic.Client,
-	qo *queryOption,
-	source *elastic.SearchSource,
-	countQuery elastic.Query,
-	queryErr error,
-	res *elastic.SearchResult,
-) (*elastic.SearchResult, bool) {
-	if client == nil || qo == nil || qo.query == nil || source == nil || countQuery == nil {
-		return nil, false
-	}
-	if !canFallbackMissingMappingQuery(qo.query) {
-		return nil, false
-	}
-
-	failures := missingMappingSortFailures(queryErr, res)
-	if len(failures) == 0 {
-		return nil, false
-	}
-
-	field := failures[0].Field
-	failedIndexes := dedupeMissingMappingFailureIndexes(failures)
-	span.Set("fallback_reason", "missing_mapping_empty_index")
-	span.Set("fallback_field", field)
-	span.Set("fallback_failed_indexes", failedIndexes)
-
-	indices, err := client.IndexGet(qo.indexes...).Do(ctx)
-	if err != nil {
-		span.Set("fallback_error", fmt.Sprintf("index_get: %v", err))
-		return nil, false
-	}
-
-	physicalIndexes := make([]string, 0, len(indices))
-	for index := range indices {
-		physicalIndexes = append(physicalIndexes, index)
-	}
-	sort.Strings(physicalIndexes)
-	if len(physicalIndexes) == 0 {
-		span.Set("fallback_error", "empty_resolved_indexes")
-		return nil, false
-	}
-
-	remainingPhysicalIndexes := excludeStrings(physicalIndexes, failedIndexes)
-	if len(remainingPhysicalIndexes) == 0 {
-		span.Set("fallback_error", "empty_retry_indexes")
-		return nil, false
-	}
-
-	matchedDocs, checkErr := searchExactTotalHits(ctx, client, failedIndexes, countQuery)
-	if checkErr != nil {
-		span.Set("fallback_error", fmt.Sprintf("empty_check: %v", checkErr))
-		return nil, false
-	}
-	if matchedDocs != 0 {
-		span.Set("fallback_error", fmt.Sprintf("non_empty_indexes: count=%d", matchedDocs))
-		return nil, false
-	}
-
-	retryIndexes := appendExcludedIndexExpressions(qo.indexes, failedIndexes)
-	span.Set("fallback_retry_indexes", retryIndexes)
-
-	log.Warnf(
-		ctx,
-		"es missing mapping fallback triggered field: %s, failed_indexes: %+v, retry_indexes: %+v, matched_docs: %d",
-		field, failedIndexes, retryIndexes, matchedDocs,
-	)
-
-	retryRes, retryErr := client.Search().Index(retryIndexes...).SearchSource(source).Do(ctx)
-	if retryErr != nil {
-		span.Set("fallback_error", fmt.Sprintf("retry: %v", retryErr))
-		return nil, false
-	}
-	return retryRes, true
-}
-
-func searchExactTotalHits(ctx context.Context, client *elastic.Client, indexes []string, query elastic.Query) (int64, error) {
-	res, err := client.Search().
-		Index(indexes...).
-		Query(query).
-		Size(0).
-		TrackTotalHits(true).
-		Do(ctx)
-	if err != nil {
-		return 0, err
-	}
-	if res == nil {
-		return 0, fmt.Errorf("empty response")
-	}
-	if res.TimedOut {
-		return 0, fmt.Errorf("timed out")
-	}
-	if res.Shards == nil {
-		return 0, fmt.Errorf("shards info is nil")
-	}
-	if res.Shards.Failed != 0 {
-		return 0, fmt.Errorf(
-			"shards failed: total=%d successful=%d failed=%d",
-			res.Shards.Total, res.Shards.Successful, res.Shards.Failed,
-		)
-	}
-	if res.Hits == nil || res.Hits.TotalHits == nil {
-		return 0, fmt.Errorf("total hits is nil")
-	}
-	if res.Hits.TotalHits.Relation != "eq" {
-		return 0, fmt.Errorf("total hits is not exact: relation=%s value=%d", res.Hits.TotalHits.Relation, res.Hits.TotalHits.Value)
-	}
-	return res.Hits.TotalHits.Value, nil
-}
-
-func canFallbackMissingMappingQuery(query *metadata.Query) bool {
-	if query == nil {
-		return false
-	}
-	if query.Scroll != "" {
-		return false
-	}
-	if query.ResultTableOption == nil {
-		return true
-	}
-	if query.ResultTableOption.ScrollID != "" {
-		return false
-	}
-	return len(query.ResultTableOption.SearchAfter) == 0
-}
-
-func missingMappingSortFailures(err error, res *elastic.SearchResult) []esMissingMappingFailure {
-	shardFailures, extractedErr := extractESResult(err, res)
-	failures := make([]esMissingMappingFailure, 0, len(shardFailures))
-	for _, failure := range shardFailures {
-		if failure == nil || failure.Index == "" || failure.Reason == nil {
-			continue
-		}
-		failures = appendMissingMappingFailure(failures, failure.Index, failure.Reason)
-	}
-
-	var esErr *elastic.Error
-	if errors.As(extractedErr, &esErr) && esErr != nil && esErr.Details != nil {
-		for _, failedShard := range esErr.Details.FailedShards {
-			index, _ := failedShard[IndexField].(string)
-			reason, _ := failedShard[ReasonField].(map[string]any)
-			failures = appendMissingMappingFailure(failures, index, reason)
-		}
-	}
-	return failures
-}
-
-func appendMissingMappingFailure(failures []esMissingMappingFailure, index string, reason map[string]any) []esMissingMappingFailure {
-	if index == "" || reason == nil {
-		return failures
-	}
-	reasonMsg, _ := extractReasonAndType(reason, true)
-	if reasonMsg == "" {
-		reasonMsg, _ = extractReasonAndType(reason, false)
-	}
-	field, ok := missingMappingSortField(reasonMsg)
-	if !ok {
-		return failures
-	}
-	return append(failures, esMissingMappingFailure{
-		Index: index,
-		Field: field,
-	})
-}
-
-func missingMappingSortField(reason string) (string, bool) {
-	if !strings.HasPrefix(reason, esMissingMappingReasonPrefix) || !strings.HasSuffix(reason, esMissingMappingReasonSuffix) {
-		return "", false
-	}
-	field := strings.TrimSuffix(strings.TrimPrefix(reason, esMissingMappingReasonPrefix), esMissingMappingReasonSuffix)
-	if field == "" {
-		return "", false
-	}
-	return field, true
-}
-
-func dedupeMissingMappingFailureIndexes(failures []esMissingMappingFailure) []string {
-	indexSet := make(map[string]struct{}, len(failures))
-	for _, failure := range failures {
-		if failure.Index == "" {
-			continue
-		}
-		indexSet[failure.Index] = struct{}{}
-	}
-	return sortedStringSetKeys(indexSet)
-}
-
-func makeStringSet(values []string) map[string]struct{} {
-	set := make(map[string]struct{}, len(values))
-	for _, value := range values {
-		set[value] = struct{}{}
-	}
-	return set
-}
-
-func excludeStrings(values []string, excluded []string) []string {
-	excludedSet := makeStringSet(excluded)
-	filtered := make([]string, 0, len(values))
-	for _, value := range values {
-		if _, ok := excludedSet[value]; ok {
-			continue
-		}
-		filtered = append(filtered, value)
-	}
-	return filtered
-}
-
-func appendExcludedIndexExpressions(indexes []string, excluded []string) []string {
-	result := append([]string{}, indexes...)
-	for _, index := range excluded {
-		result = append(result, "-"+index)
-	}
-	return result
-}
-
-func sortedStringSetKeys(set map[string]struct{}) []string {
-	values := make([]string, 0, len(set))
-	for value := range set {
-		values = append(values, value)
-	}
-	sort.Strings(values)
-	return values
-}
-
-func esQueryIndexes(qo *queryOption) []string {
-	if qo == nil {
-		return nil
-	}
-	return qo.indexes
-}
-
-func countESShardFailures(failures []*elastic.ShardOperationFailedException) int {
-	count := 0
-	for _, failure := range failures {
-		if failure != nil {
-			count++
-		}
-	}
-	return count
-}
-
-func buildESShardFailureSample(failures []*elastic.ShardOperationFailedException) []esShardFailureSample {
-	samples := make([]esShardFailureSample, 0, esShardFailureSampleLimit)
-	for _, failure := range failures {
-		if failure == nil {
-			continue
-		}
-		samples = append(samples, esShardFailureSample{
-			Shard:  failure.Shard,
-			Index:  failure.Index,
-			Status: failure.Status,
-			Reason: truncateString(marshalESShardFailureReason(failure.Reason), esShardFailureReasonMaxLength),
-		})
-		if len(samples) >= esShardFailureSampleLimit {
-			break
-		}
-	}
-	return samples
-}
-
-func marshalESShardFailureReason(reason map[string]interface{}) string {
-	if len(reason) == 0 {
-		return ""
-	}
-	reasonJson, err := json.Marshal(reason)
-	if err != nil {
-		return fmt.Sprintf("%+v", reason)
-	}
-	return string(reasonJson)
-}
-
-func marshalESShardFailureSample(failuresSample []esShardFailureSample) string {
-	failuresJson, err := json.Marshal(failuresSample)
-	if err != nil {
-		return fmt.Sprintf("%+v", failuresSample)
-	}
-	return string(failuresJson)
-}
-
-func truncateString(s string, maxLength int) string {
-	if maxLength <= 0 {
-		return ""
-	}
-	runes := []rune(s)
-	if len(runes) <= maxLength {
-		return s
-	}
-	return string(runes[:maxLength])
 }
 
 func (i *Instance) queryWithAgg(ctx context.Context, qo *queryOption, fact *FormatFactory) storage.SeriesSet {
